Should not prevent user mulitple install spdy without reinputing key/cert

+++ setup-tools/install-spdy.sh
@@ -39,7 +39,7 @@ while [ ${maxTryCert} -gt 0 ]; do
   ((maxTryCert--))
 done
 
-if [[  -f ${key} ||  -f ${cert} ]]; then
+if [[  -f ${key} &&  -f ${cert} ]]; then
   echoS "Copy Key ${key} and Cert ${cert} to ${configDir}"
   cp ${key} ${SPDYSSLKeyFile}
   cp ${cert} ${SPDYSSLCertFile}
